Refactor out duplicate PTT logic into a shared controller.
