+++ advancedSettings.pro
@@ -23,7 +23,8 @@ SOURCES += src/main.cpp\
 		src/tabcontrollers/SteamVRTabController.cpp \
 		src/tabcontrollers/ReviveTabController.cpp \
 		src/tabcontrollers/UtilitiesTabController.cpp \
-		src/tabcontrollers/audiomanager/AudioManagerWindows.cpp
+		src/tabcontrollers/audiomanager/AudioManagerWindows.cpp \
+		src/tabcontrollers/PttController.cpp
 
 HEADERS  += src/overlaycontroller.h \
 		src/logging.h \
@@ -37,7 +38,8 @@ HEADERS  += src/overlaycontroller.h \
 		src/tabcontrollers/ReviveTabController.h \
 		src/tabcontrollers/UtilitiesTabController.h \
 		src/tabcontrollers/AudioManager.h \
-		src/tabcontrollers/audiomanager/AudioManagerWindows.h
+		src/tabcontrollers/audiomanager/AudioManagerWindows.h \
+		src/tabcontrollers/PttController.h
 
 INCLUDEPATH += third-party/openvr/include \
 			third-party/easylogging++
